backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-solo==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-dynamic-fields==0.4.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
priority==2.0.0
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
taskgroup==0.0.0a4
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
uritemplate==4.1.1
wsproto==1.2.0
zope.interface==6.2
